requirements.txt
pandas
argparse
fuzzywuzzy
python-Levenshtein
seaborn
matplotlib.pyplot